Initial work on final pass fb blit

+++ vk_renderer/GeomPass.cpp
@@ -81,7 +81,7 @@ void GeomPass::initCommandBuffers()
 	allocInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_ALLOCATE_INFO;
 	allocInfo.commandPool = VkEngine::getEngine().getCommandPool();
 	allocInfo.level = VK_COMMAND_BUFFER_LEVEL_PRIMARY;
-	allocInfo.commandBufferCount = (uint32_t) 1;
+	allocInfo.commandBufferCount = 1;
 
 	VK_CHECK(vkAllocateCommandBuffers(VkEngine::getEngine().getDevice(), &allocInfo, commandBuffers.data()));
